@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Library</OutputType>
-    <TargetFramework>net7.0-windows</TargetFramework>
+    <TargetFramework>net9.0-windows</TargetFramework>
     <UseWPF>true</UseWPF>
     <ProjectGuid>{9B130CC5-14FB-41FF-B310-0A95B6894C37}</ProjectGuid>
     <AppDesignerFolder>Properties</AppDesignerFolder>
@@ -45,14 +45,18 @@
                             $(OutputPath)runtimes\linux-musl-arm;
                             $(OutputPath)runtimes\linux-musl-arm64;
                             $(OutputPath)runtimes\linux-musl-x64;
+                            $(OutputPath)runtimes\linux-musl-s390x;
                             $(OutputPath)runtimes\linux-ppc64le;
                             $(OutputPath)runtimes\linux-s390x;
                             $(OutputPath)runtimes\linux-x64;
                             $(OutputPath)runtimes\linux-x86;
                             $(OutputPath)runtimes\maccatalyst-arm64;
                             $(OutputPath)runtimes\maccatalyst-x64;
+                            $(OutputPath)runtimes\osx;
                             $(OutputPath)runtimes\osx-arm64;
-                            $(OutputPath)runtimes\osx-x64"/>
+                            $(OutputPath)runtimes\osx-x64;
+                            $(OutputPath)runtimes\win-arm;
+                            $(OutputPath)runtimes\win-arm64;"/>
   </Target>
 
   <Target Name="RemoveUnnecessaryRuntimesAfterPublish" AfterTargets="Publish">
@@ -64,14 +68,18 @@
                             $(PublishDir)runtimes\linux-musl-arm;
                             $(PublishDir)runtimes\linux-musl-arm64;
                             $(PublishDir)runtimes\linux-musl-x64;
+                            $(PublishDir)runtimes\linux-musl-s390x;
                             $(PublishDir)runtimes\linux-ppc64le;
                             $(PublishDir)runtimes\linux-s390x;
                             $(PublishDir)runtimes\linux-x64;
                             $(PublishDir)runtimes\linux-x86;
                             $(PublishDir)runtimes\maccatalyst-arm64;
                             $(PublishDir)runtimes\maccatalyst-x64;
+                            $(PublishDir)runtimes\osx;
                             $(PublishDir)runtimes\osx-arm64;
-                            $(PublishDir)runtimes\osx-x64"/>
+                            $(PublishDir)runtimes\osx-x64;
+                            $(PublishDir)runtimes\win-arm;
+                            $(PublishDir)runtimes\win-arm64;"/>
   </Target>
 
   <ItemGroup>
@@ -81,7 +89,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Flow.Launcher.Infrastructure\Flow.Launcher.Infrastructure.csproj" />
     <ProjectReference Include="..\..\Flow.Launcher.Plugin\Flow.Launcher.Plugin.csproj" />
   </ItemGroup>
 
@@ -96,7 +103,10 @@
 
   <ItemGroup>
     <PackageReference Include="CommunityToolkit.Mvvm" Version="8.4.0" />
-    <PackageReference Include="Microsoft.Data.Sqlite" Version="9.0.3" />
+    <PackageReference Include="Flow.Launcher.Localization" Version="0.0.4" />
+    <PackageReference Include="Microsoft.Data.Sqlite" Version="9.0.8" />
+    <PackageReference Include="Svg.Skia" Version="3.0.4" />
+    <PackageReference Include="SkiaSharp" Version="3.119.0" />
   </ItemGroup>
 
 </Project>
