@@ -1,11 +1,12 @@
-﻿using Flow.Launcher.Infrastructure.Hotkey;
-using Flow.Launcher.Infrastructure.UserSettings;
-using System;
-using NHotkey;
-using NHotkey.Wpf;
-using Flow.Launcher.ViewModel;
+﻿using System;
 using ChefKeys;
 using CommunityToolkit.Mvvm.DependencyInjection;
+using Flow.Launcher.Infrastructure.Hotkey;
+using Flow.Launcher.Infrastructure.DialogJump;
+using Flow.Launcher.Infrastructure.UserSettings;
+using Flow.Launcher.ViewModel;
+using NHotkey;
+using NHotkey.Wpf;
 
 namespace Flow.Launcher.Helper;
 
@@ -22,6 +23,10 @@ internal static class HotKeyMapper
         _settings = Ioc.Default.GetService<Settings>();
 
         SetHotkey(_settings.Hotkey, OnToggleHotkey);
+        if (_settings.EnableDialogJump)
+        {
+            SetHotkey(_settings.DialogJumpHotkey, DialogJump.OnToggleHotkey);
+        }
         LoadCustomPluginHotkey();
     }
 
