@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Collections.Specialized;
@@ -8,6 +8,7 @@ using System.IO;
 using System.Linq;
 using System.Net;
 using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
@@ -74,7 +75,7 @@ namespace Flow.Launcher
         }
 
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Usage", "VSTHRD100:Avoid async void methods", Justification = "<Pending>")]
-        public async void RestartApp()
+        public void RestartApp()
         {
             _mainVM.Hide();
 
@@ -82,9 +83,6 @@ namespace Flow.Launcher
             // UpdateManager.RestartApp() will call Environment.Exit(0)
             // which will cause ungraceful exit
             SaveAppAllSettings();
-
-            // Wait for all image caches to be saved before restarting
-            await ImageLoader.WaitSaveAsync();
 
             // Restart requires Squirrel's Update.exe to be present in the parent folder, 
             // it is only published from the project's release pipeline. When debugging without it,
@@ -115,8 +113,8 @@ namespace Flow.Launcher
                 _settings.Save();
                 PluginManager.Save();
                 _mainVM.Save();
+                ImageLoader.Save();
             }
-            _ = ImageLoader.SaveAsync();
         }
 
         public Task ReloadAllPluginData() => PluginManager.ReloadDataAsync();
@@ -379,24 +377,35 @@ namespace Flow.Launcher
                     explorer.Start();
                 }
             }
+            catch (COMException ex) when (ex.ErrorCode == unchecked((int)0x80004004))
+            {
+                /*
+                 * The COMException with HResult 0x80004004 is E_ABORT (operation aborted).
+                 * Shell APIs often return this when the operation is canceled or the shell cannot complete it cleanly.
+                 * It most likely comes from Win32Helper.OpenFolderAndSelectFile(targetPath).
+                 * Typical triggers:
+                 * The target file/folder was deleted/moved between computing targetPath and the shell call.
+                 * The folder is on an offline network/removable drive.
+                 * Explorer is restarting/busy and aborts the request.
+                 * A selection request to a new/closing Explorer window is canceled.
+                 * Because it is commonly user- or environment-driven and not actionable,
+                 * we should treat it as expected noise and ignore it to avoid bothering users.
+                 */
+            }
             catch (Win32Exception ex) when (ex.NativeErrorCode == 2)
             {
                 LogError(ClassName, "File Manager not found");
-                ShowMsgBox(
-                    string.Format(GetTranslation("fileManagerNotFound"), ex.Message),
+                ShowMsgError(
                     GetTranslation("fileManagerNotFoundTitle"),
-                    MessageBoxButton.OK,
-                    MessageBoxImage.Error
+                    string.Format(GetTranslation("fileManagerNotFound"), ex.Message)
                 );
             }
             catch (Exception ex)
             {
                 LogException(ClassName, "Failed to open folder", ex);
-                ShowMsgBox(
-                    string.Format(GetTranslation("folderOpenError"), ex.Message),
+                ShowMsgError(
                     GetTranslation("errorTitle"),
-                    MessageBoxButton.OK,
-                    MessageBoxImage.Error
+                    string.Format(GetTranslation("folderOpenError"), ex.Message)
                 );
             }
         }
@@ -424,11 +433,9 @@ namespace Flow.Launcher
                 {
                     var tabOrWindow = browserInfo.OpenInTab ? "tab" : "window";
                     LogException(ClassName, $"Failed to open URL in browser {tabOrWindow}: {path}, {inPrivate ?? browserInfo.EnablePrivate}, {browserInfo.PrivateArg}", e);
-                    ShowMsgBox(
-                        GetTranslation("browserOpenError"),
+                    ShowMsgError(
                         GetTranslation("errorTitle"),
-                        MessageBoxButton.OK,
-                        MessageBoxImage.Error
+                        GetTranslation("browserOpenError")
                     );
                 }
             }
@@ -598,6 +605,10 @@ namespace Flow.Launcher
             add => _mainVM.ActualApplicationThemeChanged += value;
             remove => _mainVM.ActualApplicationThemeChanged -= value;
         }
+
+        public string GetDataDirectory() => DataLocation.DataDirectory();
+
+        public string GetLogDirectory() => DataLocation.VersionLogDirectory;
 
         #endregion
 
