@@ -7,10 +7,10 @@ using System.Windows.Interop;
 using CommunityToolkit.Mvvm.DependencyInjection;
 using Flow.Launcher.Infrastructure;
 using Flow.Launcher.Infrastructure.UserSettings;
+using Flow.Launcher.Plugin.SharedModels;
 using Flow.Launcher.SettingPages.Views;
 using Flow.Launcher.ViewModel;
 using ModernWpf.Controls;
-using Screen = System.Windows.Forms.Screen;
 
 namespace Flow.Launcher;
 
@@ -202,7 +202,7 @@ public partial class SettingWindow
 
     private static bool IsPositionValid(double top, double left)
     {
-        foreach (var screen in Screen.AllScreens)
+        foreach (var screen in MonitorInfo.GetDisplayMonitors())
         {
             var workingArea = screen.WorkingArea;
 
@@ -217,7 +217,7 @@ public partial class SettingWindow
 
     private double WindowLeft()
     {
-        var screen = Screen.FromPoint(System.Windows.Forms.Cursor.Position);
+        var screen = MonitorInfo.GetCursorDisplayMonitor();
         var dip1 = Win32Helper.TransformPixelsToDIP(this, screen.WorkingArea.X, 0);
         var dip2 = Win32Helper.TransformPixelsToDIP(this, screen.WorkingArea.Width, 0);
         var left = (dip2.X - ActualWidth) / 2 + dip1.X;
@@ -226,7 +226,7 @@ public partial class SettingWindow
 
     private double WindowTop()
     {
-        var screen = Screen.FromPoint(System.Windows.Forms.Cursor.Position);
+        var screen = MonitorInfo.GetCursorDisplayMonitor();
         var dip1 = Win32Helper.TransformPixelsToDIP(this, 0, screen.WorkingArea.Y);
         var dip2 = Win32Helper.TransformPixelsToDIP(this, 0, screen.WorkingArea.Height);
         var top = (dip2.Y - ActualHeight) / 2 + dip1.Y - 20;
