@@ -19,7 +19,7 @@ public class Main : ISettingProvider, IPlugin, IReloadable, IPluginI18n, IContex
 
     internal static string _faviconCacheDir;
 
-    internal static PluginInitContext _context;
+    internal static PluginInitContext Context { get; set; }
 
     internal static Settings _settings;
 
@@ -29,20 +29,40 @@ public class Main : ISettingProvider, IPlugin, IReloadable, IPluginI18n, IContex
     
     public void Init(PluginInitContext context)
     {
-        _context = context;
+        Context = context;
 
         _settings = context.API.LoadSettingJsonStorage<Settings>();
 
         _faviconCacheDir = Path.Combine(
             context.CurrentPluginMetadata.PluginCacheDirectoryPath,
             "FaviconCache");
+        
+        try
+        {
+            if (Directory.Exists(_faviconCacheDir))
+            {
+                var files = Directory.GetFiles(_faviconCacheDir);
+                foreach (var file in files)
+                {
+                    var extension = Path.GetExtension(file);
+                    if (extension is ".db-shm" or ".db-wal" or ".sqlite-shm" or ".sqlite-wal")
+                    {
+                        File.Delete(file);
+                    }
+                }
+            }
+        }
+        catch (Exception e)
+        {
+            Context.API.LogException(ClassName, "Failed to clean up orphaned cache files.", e);
+        }
 
         LoadBookmarksIfEnabled();
     }
 
     private static void LoadBookmarksIfEnabled()
     {
-        if (_context.CurrentPluginMetadata.Disabled)
+        if (Context.CurrentPluginMetadata.Disabled)
         {
             // Don't load or monitor files if disabled
             return;
@@ -84,7 +104,7 @@ public class Main : ISettingProvider, IPlugin, IReloadable, IPluginI18n, IContex
                         Score = BookmarkLoader.MatchProgram(c, param).Score,
                         Action = _ =>
                         {
-                            _context.API.OpenUrl(c.Url);
+                            Context.API.OpenUrl(c.Url);
 
                             return true;
                         },
@@ -108,7 +128,7 @@ public class Main : ISettingProvider, IPlugin, IReloadable, IPluginI18n, IContex
                         Score = 5,
                         Action = _ =>
                         {
-                            _context.API.OpenUrl(c.Url);
+                            Context.API.OpenUrl(c.Url);
                             return true;
                         },
                         ContextData = new BookmarkAttributes { Url = c.Url }
@@ -192,12 +212,12 @@ public class Main : ISettingProvider, IPlugin, IReloadable, IPluginI18n, IContex
 
     public string GetTranslatedPluginTitle()
     {
-        return _context.API.GetTranslation("flowlauncher_plugin_browserbookmark_plugin_name");
+        return Localize.flowlauncher_plugin_browserbookmark_plugin_name();
     }
 
     public string GetTranslatedPluginDescription()
     {
-        return _context.API.GetTranslation("flowlauncher_plugin_browserbookmark_plugin_description");
+        return Localize.flowlauncher_plugin_browserbookmark_plugin_description();
     }
 
     public Control CreateSettingPanel()
@@ -211,23 +231,20 @@ public class Main : ISettingProvider, IPlugin, IReloadable, IPluginI18n, IContex
         {
             new()
             {
-                Title = _context.API.GetTranslation("flowlauncher_plugin_browserbookmark_copyurl_title"),
-                SubTitle = _context.API.GetTranslation("flowlauncher_plugin_browserbookmark_copyurl_subtitle"),
+                Title = Localize.flowlauncher_plugin_browserbookmark_copyurl_title(),
+                SubTitle = Localize.flowlauncher_plugin_browserbookmark_copyurl_subtitle(),
                 Action = _ =>
                 {
                     try
                     {
-                        _context.API.CopyToClipboard(((BookmarkAttributes)selectedResult.ContextData).Url);
+                        Context.API.CopyToClipboard(((BookmarkAttributes)selectedResult.ContextData).Url);
 
                         return true;
                     }
                     catch (Exception e)
                     {
-                        var message = "Failed to set url in clipboard";
-                        _context.API.LogException(ClassName, message, e);
-
-                        _context.API.ShowMsg(message);
-
+                        Context.API.LogException(ClassName, "Failed to set url in clipboard", e);
+                        Context.API.ShowMsgError(Localize.flowlauncher_plugin_browserbookmark_copy_failed());
                         return false;
                     }
                 },
