@@ -1,18 +1,17 @@
-﻿using System.ComponentModel;
-using Flow.Launcher.Core.Resource;
+﻿using Flow.Launcher.Localization.Attributes;
 
 namespace Flow.Launcher.Plugin.Calculator
 {
-    [TypeConverter(typeof(LocalizationConverter))]
+    [EnumLocalize]
     public enum DecimalSeparator
     {
-        [LocalizedDescription("flowlauncher_plugin_calculator_decimal_seperator_use_system_locale")]
+        [EnumLocalizeKey(nameof(Localize.flowlauncher_plugin_calculator_decimal_separator_use_system_locale))]
         UseSystemLocale,
         
-        [LocalizedDescription("flowlauncher_plugin_calculator_decimal_seperator_dot")]
+        [EnumLocalizeKey(nameof(Localize.flowlauncher_plugin_calculator_decimal_separator_dot))]
         Dot, 
         
-        [LocalizedDescription("flowlauncher_plugin_calculator_decimal_seperator_comma")]
+        [EnumLocalizeKey(nameof(Localize.flowlauncher_plugin_calculator_decimal_separator_comma))]
         Comma
     }
 }
