@@ -10,10 +10,12 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Controls;
 using Flow.Launcher.Plugin.Explorer.Exceptions;
+using System.Linq;
+using System.Globalization;
 
 namespace Flow.Launcher.Plugin.Explorer
 {
-    public class Main : ISettingProvider, IAsyncPlugin, IContextMenu, IPluginI18n
+    public class Main : ISettingProvider, IAsyncPlugin, IContextMenu, IPluginI18n, IAsyncDialogJump
     {
         internal static PluginInitContext Context { get; set; }
 
@@ -21,9 +23,11 @@ namespace Flow.Launcher.Plugin.Explorer
 
         private SettingsViewModel viewModel;
 
-        private IContextMenu contextMenu;
+        private ContextMenu contextMenu;
 
         private SearchManager searchManager;
+
+        private static readonly List<DialogJumpResult> _emptyDialogJumpResultList = new();
 
         public Control CreateSettingPanel()
         {
@@ -38,12 +42,9 @@ namespace Flow.Launcher.Plugin.Explorer
             FillQuickAccessLinkNames();
 
             viewModel = new SettingsViewModel(context, Settings);
-
-            contextMenu = new ContextMenu(Context, Settings, viewModel);
+            contextMenu = new ContextMenu(Context, Settings);
             searchManager = new SearchManager(Settings, Context);
             ResultManager.Init(Context, Settings);
-            
-            SortOptionTranslationHelper.API = context.API;
 
             EverythingApiDllImport.Load(Path.Combine(Context.CurrentPluginMetadata.PluginDirectory, "EverythingSDK",
                 Environment.Is64BitProcess ? "x64" : "x86"));
@@ -97,6 +98,12 @@ namespace Flow.Launcher.Plugin.Explorer
             return Context.API.GetTranslation("plugin_explorer_plugin_description");
         }
 
+        public void OnCultureInfoChanged(CultureInfo newCulture)
+        {
+            // Update labels for setting view model
+            EverythingSortOptionLocalized.UpdateLabels(viewModel.AllEverythingSortOptions);
+        }
+
         private static void FillQuickAccessLinkNames()
         {
             // Legacy version does not have names for quick access links, so we fill them with the path name.
@@ -106,6 +113,19 @@ namespace Flow.Launcher.Plugin.Explorer
                 {
                     link.Name = link.Path.GetPathName();
                 }
+            }
+        }
+
+        public async Task<List<DialogJumpResult>> QueryDialogJumpAsync(Query query, CancellationToken token)
+        {
+            try
+            {
+                var results = await searchManager.SearchAsync(query, token);
+                return results.Select(r => DialogJumpResult.From(r, r.CopyText)).ToList();
+            }
+            catch (Exception e) when (e is SearchException or EngineNotAvailableException)
+            {
+                return _emptyDialogJumpResultList;
             }
         }
     }
