@@ -6,7 +6,6 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Controls;
-using Flow.Launcher.Infrastructure.UserSettings;
 using Flow.Launcher.Plugin.Program.Programs;
 using Flow.Launcher.Plugin.Program.Views;
 using Flow.Launcher.Plugin.Program.Views.Models;
@@ -234,13 +233,21 @@ namespace Flow.Launcher.Plugin.Program
                     }
                 }
 
-                // Move old cache files to the new cache directory
-                var oldWin32CacheFile = Path.Combine(DataLocation.CacheDirectory, $"{Win32CacheName}.cache");
-                var newWin32CacheFile = Path.Combine(pluginCacheDirectory, $"{Win32CacheName}.cache");
-                MoveFile(oldWin32CacheFile, newWin32CacheFile);
-                var oldUWPCacheFile = Path.Combine(DataLocation.CacheDirectory, $"{UwpCacheName}.cache");
-                var newUWPCacheFile = Path.Combine(pluginCacheDirectory, $"{UwpCacheName}.cache");
-                MoveFile(oldUWPCacheFile, newUWPCacheFile);
+                // If plugin cache directory is this: D:\\Data\\Cache\\Plugins\\Flow.Launcher.Plugin.Program
+                // then the parent directory is: D:\\Data\\Cache
+                // So we can use the parent of the parent directory to get the cache directory path
+                var directoryInfo = new DirectoryInfo(pluginCacheDirectory);
+                var cacheDirectory = directoryInfo.Parent?.Parent?.FullName;
+                // Move old cache files to the new cache directory if cache directory exists
+                if (!string.IsNullOrEmpty(cacheDirectory))
+                {
+                    var oldWin32CacheFile = Path.Combine(cacheDirectory, $"{Win32CacheName}.cache");
+                    var newWin32CacheFile = Path.Combine(pluginCacheDirectory, $"{Win32CacheName}.cache");
+                    MoveFile(oldWin32CacheFile, newWin32CacheFile);
+                    var oldUWPCacheFile = Path.Combine(cacheDirectory, $"{UwpCacheName}.cache");
+                    var newUWPCacheFile = Path.Combine(pluginCacheDirectory, $"{UwpCacheName}.cache");
+                    MoveFile(oldUWPCacheFile, newUWPCacheFile);
+                }
 
                 await _win32sLock.WaitAsync();
                 _win32s = await context.API.LoadCacheBinaryStorageAsync(Win32CacheName, pluginCacheDirectory, new List<Win32>());
@@ -446,7 +453,7 @@ namespace Flow.Launcher.Plugin.Program
                 var title = Context.API.GetTranslation("flowlauncher_plugin_program_disable_dlgtitle_error");
                 var message = string.Format(Context.API.GetTranslation("flowlauncher_plugin_program_run_failed"),
                     info.FileName);
-                Context.API.ShowMsg(title, string.Format(message, info.FileName), string.Empty);
+                Context.API.ShowMsgError(title, message);
             }
         }
 
