@@ -462,7 +462,7 @@ namespace Flow.Launcher.Plugin.Program.Programs
                         var message =
                             api.GetTranslation(
                                 "flowlauncher_plugin_program_run_as_administrator_not_supported_message");
-                        api.ShowMsg(title, message, string.Empty);
+                        api.ShowMsgError(title, message);
                     }
 
                     return true;
