@@ -70,7 +70,7 @@ namespace Flow.Launcher.Plugin.Url
                             }
                             catch(Exception)
                             {
-                                context.API.ShowMsg(string.Format(context.API.GetTranslation("flowlauncher_plugin_url_cannot_open_url"), raw));
+                                context.API.ShowMsgError(string.Format(context.API.GetTranslation("flowlauncher_plugin_url_cannot_open_url"), raw));
                                 return false;
                             }
                         }
