@@ -7,6 +7,7 @@ namespace Flow.Launcher.Plugin.WebSearch
     public class SearchSource : BaseModel
     {
         public string Title { get; set; }
+
         public string ActionKeyword { get; set; }
 
         [NotNull]
@@ -19,21 +20,17 @@ namespace Flow.Launcher.Plugin.WebSearch
         /// Custom icons are placed in the user data directory
         /// </summary>
         [JsonIgnore]
-        public string IconPath 
-        {
-            get
-            {
-                if (CustomIcon)
-                    return Path.Combine(Main.CustomImagesDirectory, Icon);
-
-                return Path.Combine(Main.DefaultImagesDirectory, Icon);
-            }
-        }
+        public string IconPath => CustomIcon 
+            ? Path.Combine(Main.CustomImagesDirectory, Icon) 
+            : Path.Combine(Main.DefaultImagesDirectory, Icon);
 
         public string Url { get; set; }
 
         [JsonIgnore]
         public bool Status => Enabled;
+
+        public bool IsPrivateMode { get; set; }
+
         public bool Enabled { get; set; }
 
         public SearchSource DeepCopy()
@@ -45,8 +42,10 @@ namespace Flow.Launcher.Plugin.WebSearch
                 Url = Url,
                 Icon = Icon,
                 CustomIcon = CustomIcon,
+                IsPrivateMode = IsPrivateMode,
                 Enabled = Enabled
             };
+
             return webSearch;
         }
     }
